Add duplicate deletion & point out potential issue
Use list for each safe to traverse and delete duplicate elements.
If safe points to head, an additional condition is used to prevent
accessing head's value, which might trigger a NULL access error.

Potential issue
If the function uses ascend sorting, it will not work.
The reason is still under investigation.

Change-Id: I44127c3aad6514025e06e819e6778c9dbe6deb64