Import Simplified Merge Sort
Removed unnecessary attributes to simplify the merge sort function.
Added a prefix to make the function more identifiable for the original
sort implementation in the project.

Also added the merge sort to the queue test file.

This change makes the code easier to understand and maintain by reducing
complexity without affecting the core functionality.

Change-Id: Iae3ca4258b33b32e0670db70ad8287cf063f89bc